requirements.txt
Django==1.7
MySQL-python==1.2.5
PyJWT==1.3.0
PyMySQL==0.6.6
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
django-autocomplete-light==2.0.9
django-celery==3.1.16
django-debug-toolbar==1.2.2
django-grappelli==2.6.4
django-taggit==0.16.1
kombu==3.0.26
oauthlib==0.7.2
python-openid==2.2.5
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sqlparse==0.1.15
wsgiref==0.1.2
embedly==0.5.0
nltk==3.0.3
python-social-auth==0.2.12
beautifulsoup4==4.3.2
